move views components to view directory
