+++ src/app/views/vegetable-list-view/vegetable-list-view.component.css
